Angus/region list

+++ src/stores/RegionSetStore.ts
@@ -22,10 +22,10 @@ export class RegionSetStore {
         this.frame = frame;
         this.backendService = backendService;
         this.regions = [];
-        this.selectedRegion = null;
         this.newRegionType = CARTA.RegionType.RECTANGLE;
         this.mode = RegionMode.MOVING;
         this.addPointRegion({x: 0, y: 0}, true);
+        this.selectedRegion = this.regions[0];
     }
 
     // temporary region IDs are < 0 and used
@@ -92,14 +92,21 @@ export class RegionSetStore {
         }
     };
 
+    @action selectRegionByIndex = (index: number) => {
+        if (index >= 0 && index < this.regions.length) {
+            this.selectedRegion = this.regions[index];
+        }
+    };
+
     @action deselectRegion = () => {
         this.selectedRegion = null;
     };
 
     @action deleteRegion = (region: RegionStore) => {
-        if (region && this.regions.length) {
+        // Cursor region cannot be deleted
+        if (region && region.regionId !== 0 && this.regions.length) {
             if (region === this.selectedRegion) {
-                this.selectedRegion = null;
+                this.selectedRegion = this.regions[0];
             }
             this.regions = this.regions.filter(r => r !== region);
             this.backendService.removeRegion(region.regionId);
